+++ src/Core/OOOpenGLMatrixManager.h
@@ -56,7 +56,7 @@ enum
 - (void) dealloc;
 - (void) push: (OOMatrix) matrix;
 - (OOMatrix) pop;
-- (unsigned int) stackCount;
+- (NSUInteger) stackCount;
 
 @end
 
@@ -81,7 +81,7 @@ enum
 - (void) pushModelView;
 - (OOMatrix) popModelView;
 - (OOMatrix) getModelView;
-- (unsigned int) countModelView;
+- (NSUInteger) countModelView;
 - (void) syncModelView;
 - (void) loadProjection: (OOMatrix) matrix;
 - (void) multProjection: (OOMatrix) matrix;
